ci: update test automation to run PG17 (#72)
* fix: Reorganize build ordering due to pg_config

Upstream nixpkgs separated out pg_config into a separate development
tool package, rather than being part of the std. PostgreSQL package.
However, installing a PG extension is only part of the std PostgreSQL
package, so we can't get a pg_config that references a PG with
extensions preconfigured.  To fix this, reorder the package builds so
that we build up from the Python module first and assemble the
PostgreSQL package last.

* fix: enable PG17 tests and fix initdb failures

* test: add addt sanity-check that all tests are running